Restarting the project using the hypermodern python template.
Added a few items to the documentation as well.

The .rpv files are UO Razor packet recordings of Orcish raids. I need to
get the UO client/razor assistant version pair that will allow for the
packets to be parsed correctly and I'm not sure if the files indicate
the version.