Attempt to add loop to an overarching game object.
This didn't work because of how the asyncio event loop is running.
Instead, what happens is that the game starts up and we get into the
event loop and the first second we are supposed to await something the
process exits out.

I believe that there may be a way to fix it so that the current logic
would work with the game object owning the event loop but I'm fairly
certain the easier-to-implement option is to instead have the event loop
initialize the game object instead and keep the event loop as an async
function.