Got Pygame working!
I did a little more than that, but this definitely works for now.

I am having issues however with the async eventloop not wanting to be a
method in a game object and instead have chosen to instantiate a game
object and have it as a parameter to the async loop function for now
because I know I can get it working.

In the future, I will need to look into what the best practices for the
event loop are and ensure that I am following them. I had thought
something like having it as a `play` method on the game object made the
most sense but it absolutely didn't like that. I tried using
`asyncio.run` at all of the different levels with no luck.